MAPeD-SMD: mpd_ARR_SCREEN.count u32->u16

+++ samples/smd/common/mpd_common.h
@@ -50,7 +50,7 @@ typedef struct
 
 typedef struct
 {
-	u32			count;
+	u16			count;
 	struct mpd_SCREEN**	data;
 } mpd_ARR_SCREEN;
